pyproject.toml
[tool.poetry]
name = "moneybuddyvenv"
version = "0.1.0"
description = "MoneyBuddy API - Personal budget management system"
authors = ["alessiogandelli <alessiogandelli99@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
python-dotenv = "^1.0.0"
flask = "^3.1.0"
flask-cors = "^6.0.1"




[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
 